requirements.txt
google.cloud.storage
flask
gunicorn
gevent